+++ backend/canisters/registry/api/src/lib.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Display, Formatter};
-use types::{CanisterId, TimestampMillis};
+use types::{CanisterId, Milliseconds, TimestampMillis};
 
 mod lifecycle;
 mod queries;
@@ -22,7 +22,7 @@ pub struct TokenDetails {
     pub decimals: u8,
     pub fee: u128,
     pub logo: String,
-    pub nervous_system: Option<NervousSystem>,
+    pub nervous_system: Option<NervousSystemDeprecated>,
     pub info_url: String,
     pub how_to_buy_url: String,
     pub transaction_url_format: String,
@@ -30,11 +30,65 @@ pub struct TokenDetails {
     pub last_updated: TimestampMillis,
 }
 
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct NervousSystemDetails {
+    pub root_canister_id: CanisterId,
+    pub governance_canister_id: CanisterId,
+    pub swap_canister_id: CanisterId,
+    pub ledger_canister_id: CanisterId,
+    pub index_canister_id: CanisterId,
+    pub name: String,
+    pub url: Option<String>,
+    pub logo: String,
+    pub description: Option<String>,
+    pub transaction_fee: u64,
+    pub min_neuron_stake: u64,
+    pub min_dissolve_delay_to_vote: Milliseconds,
+    pub proposal_rejection_fee: u64,
+    pub is_nns: bool,
+    #[serde(default)]
+    pub submitting_proposals_enabled: bool,
+    pub added: TimestampMillis,
+    pub last_updated: TimestampMillis,
+}
+
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-pub struct NervousSystem {
+pub struct NervousSystemSummary {
+    pub governance_canister_id: CanisterId,
+    pub ledger_canister_id: CanisterId,
+    pub is_nns: bool,
+    pub proposal_rejection_fee: u64,
+    pub submitting_proposals_enabled: bool,
+}
+
+impl From<&NervousSystemDetails> for NervousSystemSummary {
+    fn from(value: &NervousSystemDetails) -> Self {
+        NervousSystemSummary {
+            governance_canister_id: value.governance_canister_id,
+            ledger_canister_id: value.ledger_canister_id,
+            is_nns: value.is_nns,
+            proposal_rejection_fee: value.proposal_rejection_fee,
+            submitting_proposals_enabled: value.submitting_proposals_enabled,
+        }
+    }
+}
+
+// TODO Remove this after next release
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct NervousSystemDeprecated {
     pub is_nns: bool,
     pub root: CanisterId,
     pub governance: CanisterId,
+}
+
+impl From<NervousSystemDetails> for NervousSystemDeprecated {
+    fn from(value: NervousSystemDetails) -> Self {
+        NervousSystemDeprecated {
+            is_nns: value.is_nns,
+            root: value.root_canister_id,
+            governance: value.governance_canister_id,
+        }
+    }
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]