+++ backend/canisters/registry/impl/src/updates/mod.rs
@@ -1,3 +1,4 @@
 pub mod add_token;
+pub mod c2c_set_submitting_proposals_enabled;
 pub mod update_token;
 pub mod wallet_receive;