+++ frontend/openchat-agent/src/services/group/candid/idl.js
@@ -455,6 +455,7 @@ export const idlFactory = ({ IDL }) => {
     'caption' : IDL.Opt(IDL.Text),
     'prizes' : IDL.Vec(Tokens),
     'transfer' : CryptoTransaction,
+    'diamond_only' : IDL.Bool,
   });
   const MessageContentInitial = IDL.Variant({
     'Giphy' : GiphyContent,
@@ -629,6 +630,10 @@ export const idlFactory = ({ IDL }) => {
     'new_name' : IDL.Text,
     'previous_name' : IDL.Text,
   });
+  const VerifiedCredentialGate = IDL.Record({
+    'credential' : IDL.Text,
+    'issuer' : IDL.Text,
+  });
   const Milliseconds = IDL.Nat64;
   const SnsNeuronGate = IDL.Record({
     'min_stake_e8s' : IDL.Opt(IDL.Nat64),
@@ -636,6 +641,7 @@ export const idlFactory = ({ IDL }) => {
     'governance_canister_id' : CanisterId,
   });
   const AccessGate = IDL.Variant({
+    'VerifiedCredential' : VerifiedCredentialGate,
     'SnsNeuron' : SnsNeuronGate,
     'DiamondMember' : IDL.Null,
   });