+++ frontend/openchat-agent/src/services/group/candid/types.d.ts
@@ -1,7 +1,8 @@
 import type { Principal } from '@dfinity/principal';
 import type { ActorMethod } from '@dfinity/agent';
 
-export type AccessGate = { 'SnsNeuron' : SnsNeuronGate } |
+export type AccessGate = { 'VerifiedCredential' : VerifiedCredentialGate } |
+  { 'SnsNeuron' : SnsNeuronGate } |
   { 'DiamondMember' : null };
 export type AccessGateUpdate = { 'NoChange' : null } |
   { 'SetToNone' : null } |
@@ -1212,6 +1213,7 @@ export interface PrizeContentInitial {
   'caption' : [] | [string],
   'prizes' : Array<Tokens>,
   'transfer' : CryptoTransaction,
+  'diamond_only' : boolean,
 }
 export interface PrizeWinnerContent {
   'transaction' : CompletedCryptoTransaction,
@@ -1611,6 +1613,10 @@ export interface UsersInvited {
 export interface UsersUnblocked {
   'user_ids' : Array<UserId>,
   'unblocked_by' : UserId,
+}
+export interface VerifiedCredentialGate {
+  'credential' : string,
+  'issuer' : string,
 }
 export type Version = number;
 export interface VersionedRules {